pyproject.toml
[project]
name = "penguin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "nextcord>=2.6.0",
    "psutil>=7.0.0",
    "pytest>=8.4.0",
    "python-dotenv>=1.1.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "setuptools>=80.9.0",
]
